+++ application/common/components/IdStoreBase.php
@@ -2,7 +2,6 @@
 namespace Sil\Idp\IdSync\common\components;
 
 use Sil\Idp\IdSync\common\components\adapters\GoogleSheetsIdStore;
-use Sil\Idp\IdSync\common\components\adapters\InsiteIdStore;
 use Sil\Idp\IdSync\common\components\adapters\SagePeopleIdStore;
 use Sil\Idp\IdSync\common\components\adapters\WorkdayIdStore;
 use Sil\Idp\IdSync\common\components\adapters\fakes\FakeIdStore;
@@ -14,14 +13,12 @@ abstract class IdStoreBase extends Component implements IdStoreInterface
 {
     const ADAPTER_FAKE = 'fake';
     const ADAPTER_GOOGLE_SHEETS = 'googlesheets';
-    const ADAPTER_INSITE = 'insite';
     const ADAPTER_WORKDAY = 'workday';
     const ADAPTER_SAGE_PEOPLE = 'sagepeople';
 
     protected static $adapters = [
         self::ADAPTER_FAKE => FakeIdStore::class,
         self::ADAPTER_GOOGLE_SHEETS => GoogleSheetsIdStore::class,
-        self::ADAPTER_INSITE => InsiteIdStore::class,
         self::ADAPTER_WORKDAY => WorkdayIdStore::class,
         self::ADAPTER_SAGE_PEOPLE => SagePeopleIdStore::class,
     ];